Mid implementation of DRAW instruction.
Seem to be running into a weird bug with the draw instruction. There's a
lot of complicated logic inside of it, and I'm unsure if the problem is
in the draw logic, or if previous instructions may be leading to
cascading errors.

Considering writing some tests for the prior instructions to ensure they
work properly. Committing now to backup the in-progress work. May squash
this commit at a later date.