Fix bugs in main chip8() loop.
chip8() was not calling sys_init and also did not store the return value
of sdl_update() anywhere. sdl_update's return value is now assigned to
quit, a flag var that is used to terminate the program. The return
values of sdl_update have also been changed to align with what is
expected in the main chip() loop.