Refactor code to guard SDL logic.
Had to declare all variables at the top of the block and hide all io
calls behind an ifdef for the SDL symbol. After making these changes the
code successfully compiles using cc65! The project is not yet in a state
where I can test the cc65 binary on the Apple ][, but successful
compilation is a promising sign!