Refactored makefile to try get cc65 to work.
Idk what's going on. This makefile seems to successfully compile
multiple c files? And then puts the .o files in obj and then makes a
binary? But when I load it onto the Apple ][ I just get a bunch of
blinking characters.

Unsure what the problem is tbh. Maybe writing to screen buffer memory
when loading the file?