Write a ruby script to generate rom.c
Kind of weird to do in the middle of all this build pipeline debugging.
This is a quick and dirty script that will read ROMs/IBM.ch8 and write
the hardcoded hex values into a function that can then load that data
into a struct for the Apple ][.

It is a very unfinished script and more a proof of concept to revisit
later.