"make disk" now uses a template disk image.
Tried a number of different disk images as part of build pipeline. The
images from apple2.org.za/ftp.apple.asimov.net/images/masters/prodos/
seem to work. I tried downloading a modern release from
https://prodos8.com/ but was not successful in getting that running.

With the 1990 disk, and mystery ProDOS_8.dsk file it seems the only
thing you need is the PRODOS sys program for things to run.

I created PRODOS8_TEMPLATE.dsk, which is a copy of ProDOS8.1990.dsk
which has had all files but PRODOS removed from it. This what is now
used to generate a disk image in the "make disk" command.